renamings and clean ups
