--- src/Microsoft.ApplicationInsights/netstandard1.3/Extensibility/Implementation/Platform/PlatformImplementation.cs
@@ -1,55 +1,0 @@
-﻿#if NETSTANDARD1_3
-namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Platform
-{
-    using System;
-    using System.Collections.Generic;
-    using Microsoft.ApplicationInsights.Extensibility.Implementation.External;
-
-    internal class PlatformImplementation : IPlatform
-    {
-        private IDebugOutput debugOutput = null;
-        
-        public IDictionary<string, object> GetApplicationSettings()
-        {
-            return null;
-        }
-
-        public string ReadConfigurationXml()
-        {
-            return null;
-        }
-
-        public ExceptionDetails GetExceptionDetails(Exception exception, ExceptionDetails parentExceptionDetails)
-        {
-            return ExceptionConverter.ConvertToExceptionDetails(exception, parentExceptionDetails);
-        }
-
-        /// <summary>
-        /// Returns the platform specific Debugger writer to the VS output console.
-        /// </summary>
-        public IDebugOutput GetDebugOutput()
-        {
-            if (this.debugOutput == null)
-            {
-                this.debugOutput = new TelemetryDebugWriter(); 
-            }
-            
-            return this.debugOutput;
-        }
-
-        public string GetEnvironmentVariable(string name)
-        {
-            return Environment.GetEnvironmentVariable(name);
-        }
-
-        /// <summary>
-        /// Returns the machine name.
-        /// </summary>
-        /// <returns>The machine name.</returns>
-        public string GetMachineName()
-        {
-            return Environment.GetEnvironmentVariable("COMPUTERNAME");
-        }
-    }
-}
-#endif
